Consolidate streaming code into same package
