+++ pkg/client/tests/BUILD
@@ -1,9 +1,10 @@
-package(default_visibility = ["//visibility:public"])
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-    "go_test",
+go_library(
+    name = "go_default_library",
+    srcs = ["doc.go"],
+    importpath = "k8s.io/kubernetes/pkg/client/tests",
+    visibility = ["//visibility:public"],
 )
 
 go_test(
@@ -20,8 +21,8 @@ go_test(
         "//pkg/api/testapi:go_default_library",
         "//pkg/apis/core:go_default_library",
         "//pkg/apis/core/install:go_default_library",
-        "//pkg/kubelet/server/portforward:go_default_library",
-        "//pkg/kubelet/server/remotecommand:go_default_library",
+        "//pkg/kubelet/server/streaming/portforward:go_default_library",
+        "//pkg/kubelet/server/streaming/remotecommand:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/fields:go_default_library",
@@ -44,12 +45,6 @@ go_test(
     ],
 )
 
-go_library(
-    name = "go_default_library",
-    srcs = ["doc.go"],
-    importpath = "k8s.io/kubernetes/pkg/client/tests",
-)
-
 filegroup(
     name = "package-srcs",
     srcs = glob(["**"]),
@@ -61,4 +56,5 @@ filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
+    visibility = ["//visibility:public"],
 )
