+++ pkg/client/tests/portfoward_test.go
@@ -34,7 +34,7 @@ import (
 	restclient "k8s.io/client-go/rest"
 	. "k8s.io/client-go/tools/portforward"
 	"k8s.io/client-go/transport/spdy"
-	"k8s.io/kubernetes/pkg/kubelet/server/portforward"
+	"k8s.io/kubernetes/pkg/kubelet/server/streaming/portforward"
 )
 
 // fakePortForwarder simulates port forwarding for testing. It implements
