+++ pkg/client/tests/remotecommand_test.go
@@ -40,7 +40,7 @@ import (
 	"k8s.io/client-go/transport/spdy"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 	api "k8s.io/kubernetes/pkg/apis/core"
-	"k8s.io/kubernetes/pkg/kubelet/server/remotecommand"
+	"k8s.io/kubernetes/pkg/kubelet/server/streaming/remotecommand"
 )
 
 type fakeExecutor struct {
