+++ pkg/kubelet/server/server_test.go
@@ -57,10 +57,10 @@ import (
 	_ "k8s.io/kubernetes/pkg/apis/core/install"
 	"k8s.io/kubernetes/pkg/kubelet/cm"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
-	"k8s.io/kubernetes/pkg/kubelet/server/portforward"
-	remotecommandserver "k8s.io/kubernetes/pkg/kubelet/server/remotecommand"
 	"k8s.io/kubernetes/pkg/kubelet/server/stats"
 	"k8s.io/kubernetes/pkg/kubelet/server/streaming"
+	"k8s.io/kubernetes/pkg/kubelet/server/streaming/portforward"
+	remotecommandserver "k8s.io/kubernetes/pkg/kubelet/server/streaming/remotecommand"
 	"k8s.io/kubernetes/pkg/volume"
 )
