+++ pkg/kubelet/server/streaming/server.go
@@ -35,8 +35,8 @@ import (
 	remotecommandconsts "k8s.io/apimachinery/pkg/util/remotecommand"
 	"k8s.io/client-go/tools/remotecommand"
 	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
-	"k8s.io/kubernetes/pkg/kubelet/server/portforward"
-	remotecommandserver "k8s.io/kubernetes/pkg/kubelet/server/remotecommand"
+	"k8s.io/kubernetes/pkg/kubelet/server/streaming/portforward"
+	remotecommandserver "k8s.io/kubernetes/pkg/kubelet/server/streaming/remotecommand"
 )
 
 // Server is the library interface to serve the stream requests.
@@ -81,12 +81,12 @@ type Config struct {
 	StreamCreationTimeout time.Duration
 
 	// The streaming protocols the server supports (understands and permits).  See
-	// k8s.io/kubernetes/pkg/kubelet/server/remotecommand/constants.go for available protocols.
+	// k8s.io/kubernetes/pkg/kubelet/server/streaming/remotecommand/constants.go for available protocols.
 	// Only used for SPDY streaming.
 	SupportedRemoteCommandProtocols []string
 
 	// The streaming protocols the server supports (understands and permits).  See
-	// k8s.io/kubernetes/pkg/kubelet/server/portforward/constants.go for available protocols.
+	// k8s.io/kubernetes/pkg/kubelet/server/streaming/portforward/constants.go for available protocols.
 	// Only used for SPDY streaming.
 	SupportedPortForwardProtocols []string
