+++ test/test_owners.csv
@@ -658,7 +658,7 @@ k8s.io/kubernetes/pkg/kubelet/prober/results,krousey,1,
 k8s.io/kubernetes/pkg/kubelet/qos,vishh,0,
 k8s.io/kubernetes/pkg/kubelet/secret,kevin-wangzefeng,1,
 k8s.io/kubernetes/pkg/kubelet/server,tallclair,0,
-k8s.io/kubernetes/pkg/kubelet/server/portforward,rkouj,0,
+k8s.io/kubernetes/pkg/kubelet/server/streaming/portforward,rkouj,0,
 k8s.io/kubernetes/pkg/kubelet/server/stats,tallclair,0,
 k8s.io/kubernetes/pkg/kubelet/server/streaming,caesarxuchao,1,
 k8s.io/kubernetes/pkg/kubelet/status,mwielgus,1,
